+++ dlup/experimental_backends/_openslide.py
@@ -6,7 +6,7 @@ import numpy as np
 
 import openslide
 from dlup import UnsupportedSlideError
-from dlup.backends.common import AbstractSlideBackend, check_if_mpp_is_isotropic
+from dlup.experimental_backends.common import AbstractSlideBackend, check_if_mpp_is_isotropic
 
 
 def open_slide(filename: os.PathLike) -> "OpenSlideSlide":
