+++ dlup/experimental_backends/_pyvips.py
@@ -7,7 +7,7 @@ import numpy as np
 import PIL.Image
 import pyvips
 
-from dlup.backends.common import AbstractSlideBackend, check_if_mpp_is_isotropic, numpy_to_pil
+from dlup.experimental_backends.common import AbstractSlideBackend, check_if_mpp_is_isotropic, numpy_to_pil
 from dlup.types import PathLike
 
 def open_slide(filename: os.PathLike) -> "PyVipsSlide":
